Refactor RestClient
