+++ network-store-client/src/main/java/com/powsybl/network/store/client/RestClient.java
@@ -6,80 +6,31 @@
  */
 package com.powsybl.network.store.client;
 
-import com.powsybl.commons.PowsyblException;
 import com.powsybl.network.store.model.IdentifiableAttributes;
 import com.powsybl.network.store.model.Resource;
-import com.powsybl.network.store.model.TopLevelDocument;
-import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.core.ParameterizedTypeReference;
-import org.springframework.http.*;
-import org.springframework.web.client.RestTemplate;
 
-import java.util.*;
+import java.util.List;
+import java.util.Optional;
 
 /**
  * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
  */
-public class RestClient {
+public interface RestClient {
 
-    private final RestTemplate restTemplate;
+    <T extends IdentifiableAttributes> void create(String url, List<Resource<T>> resources, Object... uriVariables);
 
-    public RestClient(RestTemplateBuilder restTemplateBuilder) {
-        this.restTemplate = Objects.requireNonNull(restTemplateBuilder).errorHandler(new RestTemplateResponseErrorHandler()).build();
-    }
+    <T extends IdentifiableAttributes> Optional<Resource<T>> getOne(String target, String url, Object... uriVariables);
 
-    public RestTemplate getRestTemplate() {
-        return restTemplate;
-    }
+    <T extends IdentifiableAttributes> List<Resource<T>> getAll(String target, String url, Object... uriVariables);
 
-    private <T extends IdentifiableAttributes> ResponseEntity<TopLevelDocument<T>> getDocument(String url, Object... uriVariables) {
-        return restTemplate.exchange(url,
-                HttpMethod.GET,
-                new HttpEntity<>(new HttpHeaders()),
-                new ParameterizedTypeReference<TopLevelDocument<T>>() {
-                },
-                uriVariables);
-    }
+    <T extends IdentifiableAttributes> void update(String url, Resource<T> resource, Object... uriVariables);
 
-    private static <T extends IdentifiableAttributes> TopLevelDocument<T> getBody(ResponseEntity<TopLevelDocument<T>> response) {
-        if (response.getBody() == null) {
-            throw new PowsyblException("Body is null");
-        }
-        return response.getBody();
-    }
+    <T extends IdentifiableAttributes> void updateAll(String url, List<Resource<T>> resources, Object... uriVariables);
 
-    public <T extends IdentifiableAttributes> void create(String url, List<Resource<T>> resources, Object... uriVariables) {
-        restTemplate.postForObject(url, resources, Void.class, uriVariables);
-    }
+    <E> E get(String uri, ParameterizedTypeReference<E> responseType, Object... uriVariables);
 
-    public void delete(String url, Object... uriVariables) {
-        restTemplate.delete(url, uriVariables);
-    }
+    void put(String uri, Object... uriVariables);
 
-    public <T extends IdentifiableAttributes> Optional<Resource<T>> get(String target, String url, Object... uriVariables) {
-        ResponseEntity<TopLevelDocument<T>> response = getDocument(url, uriVariables);
-        if (response.getStatusCode() == HttpStatus.OK) {
-            return Optional.of(getBody(response).getData().get(0));
-        } else if (response.getStatusCode() == HttpStatus.NOT_FOUND) {
-            return Optional.empty();
-        } else {
-            throw new PowsyblException("Fail to get " + target + ", status: " + response.getStatusCode());
-        }
-    }
-
-    public <T extends IdentifiableAttributes> List<Resource<T>> getAll(String target, String url, Object... uriVariables) {
-        ResponseEntity<TopLevelDocument<T>> response = getDocument(url, uriVariables);
-        if (response.getStatusCode() != HttpStatus.OK) {
-            throw new PowsyblException("Fail to get " + target + " list, status: " + response.getStatusCode());
-        }
-        return getBody(response).getData();
-    }
-
-    public <T extends IdentifiableAttributes> void update(String url, Resource<T> resource, Object... uriVariables) {
-        restTemplate.put(url, resource, uriVariables);
-    }
-
-    public <T extends IdentifiableAttributes> void updateAll(String url, List<Resource<T>> resources, Object... uriVariables) {
-        restTemplate.put(url, resources, uriVariables);
-    }
+    void delete(String url, Object... uriVariables);
 }
