+++ network-store-client/src/test/java/com/powsybl/network/store/client/PreloadingNetworkStoreClientTest.java
@@ -41,7 +41,7 @@ import static org.springframework.test.web.client.response.MockRestResponseCreat
  */
 @RunWith(SpringRunner.class)
 @RestClientTest(RestClient.class)
-@ContextConfiguration(classes = RestClient.class)
+@ContextConfiguration(classes = RestClientImpl.class)
 public class PreloadingNetworkStoreClientTest {
 
     @Autowired
